+++ sandbox/chat-notification-pubnub-example/facade/src/services/index.ts
@@ -1,0 +1,2 @@
+export * from './messageservice.service';
+export * from './notificationservice.service';