+++ services/authentication-service/src/__tests__/fixtures/providers/bearer-token-verifier.provider.ts
@@ -1,14 +1,24 @@
-import {Provider} from '@loopback/core';
+import {Provider} from '@loopback/context';
+import {verify} from 'jsonwebtoken';
 import {VerifyFunction} from 'loopback4-authentication';
+import {IAuthUserWithPermissions} from '../keys';
 
 export class BearerTokenVerifyProvider
   implements Provider<VerifyFunction.BearerFn>
 {
-  constructor() {}
-
   value(): VerifyFunction.BearerFn {
     return async (token: string) => {
-      return {id: 1, username: 'mayank'};
+      /*
+        Implementing a basic JWT token decryption here
+        Leaving the additional security to the consumer of this application
+
+        Suggestion: to revoke these tokens put them in redis or some in-memory
+        database.
+        Use global interceptor over this to apply that check on each api.
+      */
+      return verify(token, 'test', {
+        issuer: 'test',
+      }) as IAuthUserWithPermissions;
     };
   }
 }