+++ services/authentication-service/src/__tests__/fixtures/providers/local-password.provider.ts
@@ -4,11 +4,9 @@ import {VerifyFunction} from 'loopback4-authentication';
 export class TestPasswordVerifyProvider
   implements Provider<VerifyFunction.LocalPasswordFn>
 {
-  constructor() {}
-
   value(): VerifyFunction.LocalPasswordFn {
-    return async (token: string) => {
-      return {id: 1, username: 'mayank', password: 'pass'};
+    return async (username: string, password: string) => {
+      return {id: 1, username: 'test_user', password: 'temp123!@'};
     };
   }
 }