+++ services/feature-toggle-service/migrations/sqls/20210913104858-initial-down.sql
@@ -1,0 +1,2 @@
+delete from features where name in ('system-feature','user-feature','tenant-feature');
+delete from strategies where name in ('Tenant','User');