+++ services/payment-service/src/controllers/index.ts
@@ -1,0 +1,5 @@
+export * from './ping.controller';
+export * from './orders.controller';
+export * from './transactions.controller';
+export * from './payment-gateways.controller';
+export * from './templates.controller';