+++ services/payment-service/src/repositories/orders.repository.ts
@@ -1,0 +1,16 @@
+import {inject} from '@loopback/core';
+import {DefaultCrudRepository, juggler} from '@loopback/repository';
+import {PaymentDatasourceName} from '../keys';
+import {Orders} from '../models';
+
+export class OrdersRepository extends DefaultCrudRepository<
+  Orders,
+  typeof Orders.prototype.id
+> {
+  constructor(
+    @inject(`datasources.${PaymentDatasourceName}`)
+    dataSource: juggler.DataSource,
+  ) {
+    super(Orders, dataSource);
+  }
+}