+++ sandbox/in-mail-example/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>0.4.3 (2021-11-11)</small>
+
+* chore(deps): bump validator in /sandbox/in-mail-example (#412) ([69fb61d](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/69fb61d)), closes [#412](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/412)
+
+
+
+
+
+## <small>0.4.2 (2021-11-09)</small>
+
+**Note:** Version bump only for package @sourceloop/in-mail-example
+
+
+
+
+
+## <small>0.4.1 (2021-11-09)</small>
+
+**Note:** Version bump only for package @sourceloop/in-mail-example
+
+
+
+
+
+## 0.4.0 (2021-10-24)
+
+* feat(core): new service for feature toggle (#339) ([70d5285](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/70d5285)), closes [#339](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/339)
+* chore(deps): bump i from 0.3.6 to 0.3.7 in /sandbox/in-mail-example (#354) ([4b2905c](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/4b2905c)), closes [#354](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/354)
+* chore(deps): bump passport-oauth2 in /sandbox/in-mail-example (#377) ([78979d7](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/78979d7)), closes [#377](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/377)
+* fix(in-mail-service): issue #338 bug resolved (#347) ([d9389be](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/d9389be)), closes [#347](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/347)
+
+
+
+
+
+## 0.3.0 (2021-09-12)
+
+* feat(sandbox): example of pubnub as a provider in notification service (#279) ([e30a0ed](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/e30a0ed)), closes [#279](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/279) [#208](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/208) [#208](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/208) [#0](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/0) [#278](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/278) [#211](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/211) [#263](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/263) [#280](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/280) [#208](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/208) [#208](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/208) [#298](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/298) [#208](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/208) [#263](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/263) [#280](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/280) [#208](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/208) [#208](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/208)
+* chore(deps): bump path-parse in /sandbox/in-mail-example (#306) ([d27839a](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/d27839a)), closes [#306](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/306)
+
+
+
+
+
 ## <small>0.2.1 (2021-07-24)</small>
 
 * fix(core): upgrade loopback4-soft-delete ([f653e0c](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/f653e0c)), closes [#0](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/0)