+++ services/chat-service/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>1.1.5 (2021-11-11)</small>
+
+**Note:** Version bump only for package @sourceloop/chat-service
+
+
+
+
+
+## <small>1.1.4 (2021-11-09)</small>
+
+* refactor(authentication-service): version bump in all services for release (#418) ([0808265](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/0808265)), closes [#418](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/418) [#0](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/0)
+
+
+
+
+
+## <small>1.1.2 (2021-11-09)</small>
+
+**Note:** Version bump only for package @sourceloop/chat-service
+
+
+
+
+
+## <small>1.1.1 (2021-10-24)</small>
+
+* fix(core): fix all version problems ([7847008](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/7847008)), closes [#0](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/0)
+
+
+
+
+
+## <small>1.0.3 (2021-09-12)</small>
+
+* fix(core): missing open api and package lock files ([3bd9686](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/3bd9686)), closes [#0](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/0)
+* test(notification-service): added unit testcases (#264) ([6e650aa](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/6e650aa)), closes [#264](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/264) [#242](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/242) [#242](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/242) [#242](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/242) [#242](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/242)
+
+
+
+
+
 ## <small>1.0.2 (2021-07-24)</small>
 
 * fix(core): upgrade loopback4-soft-delete ([f653e0c](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/f653e0c)), closes [#0](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/0)